requirements.txt
discord.py
rwci.py
PyYAML
